PENDING: net: ti: prueth: Adds LRE stats support in Ethtool for HSR/PRP
This patch includes LRE statistics of HSR and PRP into ethtool statistics.

The firmware running on PRU is responsible for maintaining the LRE
statistics. Few of these statistics are common for both PRUs and placed
in shared memory. When user requests for stats ethtool collects all the
statistics for the specified interface and displays it in the console.

Reviewed-by: pmohan <pmohan@couthit.com>
Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Parvathi Pudi <parvathi@couthit.com>
(cherry picked from commit 4a72e55aac6888c88e462c9dfa4cf8d2729c046c)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>