PENDING: net: ti: prueth: Fix queue pointer sync across PRU's
As part of TX optimization, we have common queues which
are accessed by both PRU's inorder to avoid double copy.
It is very important to maintain synchronization of
pointers across both PRU's in order to avoid queue corruption
and non aligned access.
Since the queues are common for both PRU's it is mandatory to
update the pointers of both PRU's in order to maintain the sync.

Reviewed-by: pmohan <pmohan@couthit.com>
Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Basharath Hussain Khaja <basharath@couthit.com>
Signed-off-by: Parvathi Pudi <parvathi@couthit.com>
(cherry picked from commit 26482216a39c9653ea319cd1d6e609b8d1f0fc12)
Signed-off-by: Guillaume La Roque <glaroque@baylibre.com>